requirements.txt
apache-airflow==2.7.2
astronomer-cosmos==1.2.1
dbt-core==1.6.6
dbt-duckdb==1.6.2
